The following error was fixed:
- last movement was in June
- In July, the monthly settlement for June was written
- In July no movement happend
- In August, the settlement has to be written for July with the same
value as in July for June, but the method to write the June settlement
raises a ConstraintViolationException which is catched, but a
modification which is done prior to the insert and has to be undone
after the insert is not being undone because of this exception. The
"undone part" is now moved to a finally block.